openai/requirements.txt
openai
python-dotenv
jupyterlab
ipykernel
